refactor: :recycle: Reorganise types folder and rename utils folder to utility
