+++ src/utility/string.ts
