+++ src/utility/typing.ts
